Updated bowler-data, bowler-single-component & bowler-single-html JS files (working)
The bio doesn't show up when you click the "See My Bio" button because you are not setting the inner HTML of the "bowler-bio" element to the bowler's bio property. Instead you are setting it to "this.value" which won't work because the user is not inputting a value to the button tag. Instead you need to pass the "bowler.bio" property into the method and then interpolate the value of the property into the JavaScript code which will get returned to the HTML